+++ lib/src/db/managed/data_model_builder.dart
@@ -6,11 +6,8 @@ class DataModelBuilder {
   DataModelBuilder(ManagedDataModel dataModel, List<Type> instanceTypes) {
     instanceTypes.forEach((type) {
       var backingMirror = persistentTypeOfInstanceType(type);
-      var entity = new ManagedEntity(
-          dataModel,
-          tableNameFromClass(backingMirror),
-          reflectClass(type),
-          backingMirror);
+      var entity = new ManagedEntity(dataModel,
+          tableNameFromClass(backingMirror), reflectClass(type), backingMirror);
       entities[type] = entity;
       persistentTypeToEntityMap[entity.persistentType.reflectedType] = entity;
 
@@ -29,16 +26,14 @@ class DataModelBuilder {
   Map<Type, ManagedEntity> persistentTypeToEntityMap = {};
 
   String tableNameFromClass(ClassMirror typeMirror) {
-    var declaredTableNameClass = classHierarchyForClass(typeMirror)
-        .firstWhere((cm) => cm.staticMembers[#tableName] != null,
+    var declaredTableNameClass = classHierarchyForClass(typeMirror).firstWhere(
+        (cm) => cm.staticMembers[#tableName] != null,
         orElse: () => null);
 
     if (declaredTableNameClass == null) {
       return MirrorSystem.getName(typeMirror.simpleName);
     }
-    return declaredTableNameClass
-        .invoke(#tableName, [])
-        .reflectee;
+    return declaredTableNameClass.invoke(#tableName, []).reflectee;
   }
 
   ClassMirror persistentTypeOfInstanceType(Type instanceType) {
@@ -47,9 +42,9 @@ class DataModelBuilder {
             .simpleName}' is not subclass of 'ManagedObject'.");
 
     return classHierarchyForClass(reflectClass(instanceType))
-        .firstWhere((cm) =>
-            !cm.superclass.isSubtypeOf(reflectType(ManagedObject)),
-          orElse: () => throw ifNotFoundException)
+        .firstWhere(
+            (cm) => !cm.superclass.isSubtypeOf(reflectType(ManagedObject)),
+            orElse: () => throw ifNotFoundException)
         .typeArguments
         .first;
   }
@@ -59,9 +54,8 @@ class DataModelBuilder {
     var transientProperties = transientAttributesForEntity(entity);
     var persistentProperties = persistentAttributesForEntity(entity);
 
-    return [transientProperties, persistentProperties]
-        .expand((l) => l)
-        .fold({}, (map, attribute) {
+    return [transientProperties, persistentProperties].expand((l) => l).fold({},
+        (map, attribute) {
       if (map.containsKey(attribute.name)) {
         // If there is both a getter and setter declared to represent one transient property,
         // then we need to combine them here. No other reason a property would appear twice.
@@ -74,51 +68,53 @@ class DataModelBuilder {
     });
   }
 
-  Iterable<ManagedAttributeDescription> persistentAttributesForEntity(ManagedEntity entity) {
+  Iterable<ManagedAttributeDescription> persistentAttributesForEntity(
+      ManagedEntity entity) {
     return instanceVariablesFromClass(entity.persistentType)
-        .where((declaration) => !doesVariableMirrorRepresentRelationship(declaration))
+        .where((declaration) =>
+            !doesVariableMirrorRepresentRelationship(declaration))
         .map((declaration) {
-          var type = propertyTypeFromDeclaration(declaration);
-          if (type == null) {
-            throw new ManagedDataModelException.invalidType(entity, declaration.simpleName);
-          }
+      var type = propertyTypeFromDeclaration(declaration);
+      if (type == null) {
+        throw new ManagedDataModelException.invalidType(
+            entity, declaration.simpleName);
+      }
 
-          var attributes = attributeMetadataFromDeclaration(declaration);
-          var name = propertyNameFromDeclaration(declaration);
-          return new ManagedAttributeDescription(
-              entity, name, type,
-              primaryKey: attributes?.isPrimaryKey ?? false,
-              defaultValue: attributes?.defaultValue ?? null,
-              unique: attributes?.isUnique ?? false,
-              indexed: attributes?.isIndexed ?? false,
-              nullable: attributes?.isNullable ?? false,
-              includedInDefaultResultSet: !(attributes?.shouldOmitByDefault ??
-                  false),
-              autoincrement: attributes?.autoincrement ?? false);
-        });
+      var attributes = attributeMetadataFromDeclaration(declaration);
+      var name = propertyNameFromDeclaration(declaration);
+      return new ManagedAttributeDescription(entity, name, type,
+          primaryKey: attributes?.isPrimaryKey ?? false,
+          defaultValue: attributes?.defaultValue ?? null,
+          unique: attributes?.isUnique ?? false,
+          indexed: attributes?.isIndexed ?? false,
+          nullable: attributes?.isNullable ?? false,
+          includedInDefaultResultSet:
+              !(attributes?.shouldOmitByDefault ?? false),
+          autoincrement: attributes?.autoincrement ?? false);
+    });
   }
 
-  Iterable<ManagedAttributeDescription> transientAttributesForEntity(ManagedEntity entity) {
+  Iterable<ManagedAttributeDescription> transientAttributesForEntity(
+      ManagedEntity entity) {
     return entity.instanceType.declarations.values
         .where(isTransientPropertyOrAccessor)
         .map((declaration) {
-          var type = propertyTypeFromDeclaration(declaration);
-          if (type == null) {
-            throw new ManagedDataModelException.invalidType(entity, declaration.simpleName);
-          }
+      var type = propertyTypeFromDeclaration(declaration);
+      if (type == null) {
+        throw new ManagedDataModelException.invalidType(
+            entity, declaration.simpleName);
+      }
 
-          var name = propertyNameFromDeclaration(declaration);
-          var transience = transienceForProperty(declaration);
-          if (transience == null) {
-            throw new ManagedDataModelException.invalidTransient(entity, declaration.simpleName);
-          }
+      var name = propertyNameFromDeclaration(declaration);
+      var transience = transienceForProperty(declaration);
+      if (transience == null) {
+        throw new ManagedDataModelException.invalidTransient(
+            entity, declaration.simpleName);
+      }
 
-          return new ManagedAttributeDescription.transient(
-              entity,
-              name,
-              type,
-              transience);
-        });
+      return new ManagedAttributeDescription.transient(
+          entity, name, type, transience);
+    });
   }
 
   ManagedTransientAttribute transienceForProperty(DeclarationMirror property) {
@@ -129,9 +125,11 @@ class DataModelBuilder {
     var metadata = transientMetadataFromDeclaration(property);
     MethodMirror m = property as MethodMirror;
     if (m.isGetter && metadata.isAvailableAsOutput) {
-      return new ManagedTransientAttribute(availableAsOutput: true, availableAsInput: false);
+      return new ManagedTransientAttribute(
+          availableAsOutput: true, availableAsInput: false);
     } else if (m.isSetter && metadata.isAvailableAsInput) {
-      return new ManagedTransientAttribute(availableAsInput: true, availableAsOutput: false);
+      return new ManagedTransientAttribute(
+          availableAsInput: true, availableAsOutput: false);
     }
 
     return null;
@@ -142,11 +140,11 @@ class DataModelBuilder {
     return instanceVariablesFromClass(entity.persistentType)
         .where(doesVariableMirrorRepresentRelationship)
         .fold({}, (map, declaration) {
-          var key = MirrorSystem.getName(declaration.simpleName);
-          map[key] = relationshipFromProperty(entity, declaration);
+      var key = MirrorSystem.getName(declaration.simpleName);
+      map[key] = relationshipFromProperty(entity, declaration);
 
-          return map;
-        });
+      return map;
+    });
   }
 
   ManagedRelationshipDescription relationshipFromProperty(
@@ -168,10 +166,10 @@ class DataModelBuilder {
   }
 
   ManagedRelationshipDescription relationshipForForeignKeyProperty(
-      ManagedEntity owningEntity, ManagedEntity destinationEntity,
+      ManagedEntity owningEntity,
+      ManagedEntity destinationEntity,
       VariableMirror property) {
-    var relationship = relationshipMetadataFromProperty(
-        property);
+    var relationship = relationshipMetadataFromProperty(property);
 
     // Make sure the relationship parameters are valid
     if (relationship.onDelete == ManagedRelationshipDeleteRule.nullify &&
@@ -180,18 +178,17 @@ class DataModelBuilder {
           owningEntity, property.simpleName);
     }
 
-    var inverseProperty = inverseRelationshipProperty(owningEntity, destinationEntity, property);
+    var inverseProperty =
+        inverseRelationshipProperty(owningEntity, destinationEntity, property);
 
     // Make sure we didn't annotate both sides
-    if (relationshipMetadataFromProperty(inverseProperty) !=
-        null) {
-      throw new ManagedDataModelException.dualMetadata(
-          owningEntity, property.simpleName, destinationEntity,
-          inverseProperty.simpleName);
+    if (relationshipMetadataFromProperty(inverseProperty) != null) {
+      throw new ManagedDataModelException.dualMetadata(owningEntity,
+          property.simpleName, destinationEntity, inverseProperty.simpleName);
     }
 
     var columnType =
-      destinationEntity.attributes[destinationEntity.primaryKey].type;
+        destinationEntity.attributes[destinationEntity.primaryKey].type;
 
     return new ManagedRelationshipDescription(
         owningEntity,
@@ -201,20 +198,20 @@ class DataModelBuilder {
         relationship.onDelete,
         ManagedRelationshipType.belongsTo,
         inverseProperty.simpleName,
-        unique: !inverseProperty.type.isSubtypeOf(
-            reflectType(ManagedSet)),
+        unique: !inverseProperty.type.isSubtypeOf(reflectType(ManagedSet)),
         indexed: true,
         nullable: !relationship.isRequired,
         includedInDefaultResultSet: true);
   }
 
   ManagedRelationshipDescription relationshipDescriptionForHasManyOrOneProperty(
-      ManagedEntity owningEntity, ManagedEntity destinationEntity,
+      ManagedEntity owningEntity,
+      ManagedEntity destinationEntity,
       VariableMirror property) {
-    var managedRelationship = relationshipMetadataFromProperty(
-        property);
+    var managedRelationship = relationshipMetadataFromProperty(property);
 
-    var inverseProperty = inverseRelationshipProperty(owningEntity, destinationEntity, property);
+    var inverseProperty =
+        inverseRelationshipProperty(owningEntity, destinationEntity, property);
 
     var relType = ManagedRelationshipType.hasOne;
     if (property.type.isSubtypeOf(reflectType(ManagedSet))) {
@@ -238,8 +235,8 @@ class DataModelBuilder {
         includedInDefaultResultSet: false);
   }
 
-  ManagedEntity matchingEntityForProperty(ManagedEntity owningEntity,
-      VariableMirror property) {
+  ManagedEntity matchingEntityForProperty(
+      ManagedEntity owningEntity, VariableMirror property) {
     var typeMirror = property.type;
     if (property.type.isSubtypeOf(reflectType(ManagedSet))) {
       typeMirror = typeMirror.typeArguments.first;
@@ -247,8 +244,7 @@ class DataModelBuilder {
 
     var destinationEntity = entities[typeMirror.reflectedType];
     if (destinationEntity == null) {
-      var relationshipMetadata = relationshipMetadataFromProperty(
-          property);
+      var relationshipMetadata = relationshipMetadataFromProperty(property);
       if (relationshipMetadata.isDeferred) {
         // Then we can scan for a list of possible entities that extend
         // the interface.
@@ -274,24 +270,24 @@ class DataModelBuilder {
     return destinationEntity;
   }
 
-  List<VariableMirror> entityRelationshipsWithType(ManagedEntity entity, TypeMirror type) {
-    return instanceVariablesFromClass(entity.persistentType)
-        .where((p) {
-          if (p.type.isSubtypeOf(type)) {
-            return true;
-          }
+  List<VariableMirror> entityRelationshipsWithType(
+      ManagedEntity entity, TypeMirror type) {
+    return instanceVariablesFromClass(entity.persistentType).where((p) {
+      if (p.type.isSubtypeOf(type)) {
+        return true;
+      }
 
-          if (p.type.isSubtypeOf(reflectType(ManagedSet))
-          && p.type.typeArguments.first.isSubtypeOf(type)) {
-            return true;
-          }
+      if (p.type.isSubtypeOf(reflectType(ManagedSet)) &&
+          p.type.typeArguments.first.isSubtypeOf(type)) {
+        return true;
+      }
 
-          return false;
-        })
-        .toList();
+      return false;
+    }).toList();
   }
 
-  VariableMirror inverseRelationshipProperty(ManagedEntity owningEntity, ManagedEntity destinationEntity, VariableMirror property) {
+  VariableMirror inverseRelationshipProperty(ManagedEntity owningEntity,
+      ManagedEntity destinationEntity, VariableMirror property) {
     var metadata = relationshipMetadataFromProperty(property);
     if (metadata != null) {
       // Looking for the has-a side, which has an explicit inverse.
@@ -305,20 +301,26 @@ class DataModelBuilder {
 
         if (relationshipMetadataFromProperty(destinationProperty) != null) {
           throw new ManagedDataModelException.dualMetadata(
-              owningEntity, property.simpleName, destinationEntity, destinationProperty.simpleName);
+              owningEntity,
+              property.simpleName,
+              destinationEntity,
+              destinationProperty.simpleName);
         }
 
         return destinationProperty;
       } else {
         // Looking for the has-a side, but it is deferred.
-        var candidates = entityRelationshipsWithType(destinationEntity, owningEntity.instanceType);
+        var candidates = entityRelationshipsWithType(
+            destinationEntity, owningEntity.instanceType);
         if (candidates.length == 0) {
           throw new ManagedDataModelException.missingInverse(
               owningEntity, property.simpleName, destinationEntity, null);
         } else if (candidates.length > 1) {
           throw new ManagedDataModelException.duplicateInverse(
-              owningEntity, property.simpleName,
-              destinationEntity, candidates.map((v) => v.simpleName).toList());
+              owningEntity,
+              property.simpleName,
+              destinationEntity,
+              candidates.map((v) => v.simpleName).toList());
         }
 
         return candidates.first;
@@ -328,18 +330,19 @@ class DataModelBuilder {
       // If we have an explicit inverse, look for that first.
 
       var candidates =
-        instanceVariablesFromClass(destinationEntity.persistentType)
-          .where((p) => relationshipMetadataFromProperty(p) != null)
-          .toList();
+          instanceVariablesFromClass(destinationEntity.persistentType)
+              .where((p) => relationshipMetadataFromProperty(p) != null)
+              .toList();
 
       if (candidates.length == 0) {
         throw new ManagedDataModelException.missingInverse(
             owningEntity, property.simpleName, destinationEntity, null);
       }
 
-      var specificInverse = candidates
-          .firstWhere((p) =>
-            relationshipMetadataFromProperty(p).inversePropertyName == property.simpleName,
+      var specificInverse = candidates.firstWhere(
+          (p) =>
+              relationshipMetadataFromProperty(p).inversePropertyName ==
+              property.simpleName,
           orElse: () => null);
       if (specificInverse != null) {
         return specificInverse;
@@ -347,12 +350,13 @@ class DataModelBuilder {
 
       if (candidates.length > 1) {
         throw new ManagedDataModelException.duplicateInverse(
-            owningEntity, property.simpleName,
-            destinationEntity, candidates.map((v) => v.simpleName).toList());
+            owningEntity,
+            property.simpleName,
+            destinationEntity,
+            candidates.map((v) => v.simpleName).toList());
       }
 
       return candidates.first;
     }
   }
 }
-