+++ test/db/data_model_test.dart
@@ -8,7 +8,9 @@ void main() {
     ManagedDataModel dataModel;
     setUp(() {
       dataModel = new ManagedDataModel([User, Item, Manager]);
-      ManagedContext.defaultContext = new ManagedContext(dataModel, new DefaultPersistentStore());;
+      ManagedContext.defaultContext =
+          new ManagedContext(dataModel, new DefaultPersistentStore());
+      ;
     });
 
     test("Entities have appropriate types", () {
@@ -125,8 +127,9 @@ void main() {
       expect(relDesc.inverseKey, #items);
       expect(
           relDesc.inverseRelationship ==
-              dataModel.entityForType(User).relationships[
-                  MirrorSystem.getName(relDesc.inverseKey)],
+              dataModel
+                  .entityForType(User)
+                  .relationships[MirrorSystem.getName(relDesc.inverseKey)],
           true);
       expect(relDesc.deleteRule, ManagedRelationshipDeleteRule.cascade);
       expect(relDesc.destinationEntity == dataModel.entityForType(User), true);
@@ -139,8 +142,9 @@ void main() {
       expect(relDesc.inverseKey, #manager);
       expect(
           relDesc.inverseRelationship ==
-              dataModel.entityForType(User).relationships[
-                  MirrorSystem.getName(relDesc.inverseKey)],
+              dataModel
+                  .entityForType(User)
+                  .relationships[MirrorSystem.getName(relDesc.inverseKey)],
           true);
       expect(relDesc.deleteRule, ManagedRelationshipDeleteRule.nullify);
       expect(relDesc.destinationEntity == dataModel.entityForType(User), true);
@@ -152,8 +156,9 @@ void main() {
       expect(relDesc.inverseKey, #worker);
       expect(
           relDesc.inverseRelationship ==
-              dataModel.entityForType(Manager).relationships[
-                  MirrorSystem.getName(relDesc.inverseKey)],
+              dataModel
+                  .entityForType(Manager)
+                  .relationships[MirrorSystem.getName(relDesc.inverseKey)],
           true);
       expect(
           relDesc.destinationEntity == dataModel.entityForType(Manager), true);
@@ -203,14 +208,19 @@ void main() {
 
   group("Edge cases", () {
     test("Instances with two foreign keys to same object are distinct", () {
-      var model = new ManagedDataModel([DoubleRelationshipForeignKeyModel, DoubleRelationshipHasModel]);
+      var model = new ManagedDataModel(
+          [DoubleRelationshipForeignKeyModel, DoubleRelationshipHasModel]);
 
-      var isManyOf = model.entityForType(DoubleRelationshipForeignKeyModel).relationships["isManyOf"];
+      var isManyOf = model
+          .entityForType(DoubleRelationshipForeignKeyModel)
+          .relationships["isManyOf"];
       expect(isManyOf.inverseRelationship.name, "hasManyOf");
       expect(isManyOf.destinationEntity.tableName,
           model.entityForType(DoubleRelationshipHasModel).tableName);
 
-      var isOneOf = model.entityForType(DoubleRelationshipForeignKeyModel).relationships["isOneOf"];
+      var isOneOf = model
+          .entityForType(DoubleRelationshipForeignKeyModel)
+          .relationships["isOneOf"];
       expect(isOneOf.inverseRelationship.name, "hasOneOf");
       expect(isOneOf.destinationEntity.tableName,
           model.entityForType(DoubleRelationshipHasModel).tableName);
@@ -220,7 +230,8 @@ void main() {
   group("Valid data model with deferred types", () {
     test("Entities have correct properties and relationships", () {
       var dataModel = new ManagedDataModel([TotalModel, PartialReferenceModel]);
-      ManagedContext.defaultContext = new ManagedContext(dataModel, new DefaultPersistentStore());
+      ManagedContext.defaultContext =
+          new ManagedContext(dataModel, new DefaultPersistentStore());
 
       expect(dataModel.entities.length, 2);
 
@@ -233,15 +244,23 @@ void main() {
       expect(totalEntity.attributes["addedField"].name, isNotNull);
       expect(totalEntity.attributes["id"].isPrimaryKey, true);
       expect(totalEntity.attributes["field"].isIndexed, true);
-      expect(totalEntity.relationships["hasManyRelationship"].destinationEntity.tableName, referenceEntity.tableName);
-      expect(totalEntity.relationships["hasManyRelationship"].relationshipType, ManagedRelationshipType.hasMany);
+      expect(
+          totalEntity
+              .relationships["hasManyRelationship"].destinationEntity.tableName,
+          referenceEntity.tableName);
+      expect(totalEntity.relationships["hasManyRelationship"].relationshipType,
+          ManagedRelationshipType.hasMany);
 
-      expect(referenceEntity.relationships["foreignKeyColumn"].destinationEntity.tableName, totalEntity.tableName);
+      expect(
+          referenceEntity
+              .relationships["foreignKeyColumn"].destinationEntity.tableName,
+          totalEntity.tableName);
     });
 
     test("Will use tableName of base class if not declared in subclass", () {
       var dataModel = new ManagedDataModel([TotalModel, PartialReferenceModel]);
-      ManagedContext.defaultContext = new ManagedContext(dataModel, new DefaultPersistentStore());
+      ManagedContext.defaultContext =
+          new ManagedContext(dataModel, new DefaultPersistentStore());
       expect(dataModel.entityForType(TotalModel).tableName, "predefined");
     });
 
@@ -249,23 +268,38 @@ void main() {
       var dm1 = new ManagedDataModel([TotalModel, PartialReferenceModel]);
       var dm2 = new ManagedDataModel([PartialReferenceModel, TotalModel]);
       expect(dm1.entities.map((e) => e.tableName).contains("predefined"), true);
-      expect(dm1.entities.map((e) => e.tableName).contains("_PartialReferenceModel"), true);
+      expect(
+          dm1.entities
+              .map((e) => e.tableName)
+              .contains("_PartialReferenceModel"),
+          true);
       expect(dm2.entities.map((e) => e.tableName).contains("predefined"), true);
-      expect(dm2.entities.map((e) => e.tableName).contains("_PartialReferenceModel"), true);
+      expect(
+          dm2.entities
+              .map((e) => e.tableName)
+              .contains("_PartialReferenceModel"),
+          true);
     });
 
-    test("Partials have defaultProperties from persistent type superclasses", () {
+    test("Partials have defaultProperties from persistent type superclasses",
+        () {
       var dataModel = new ManagedDataModel([TotalModel, PartialReferenceModel]);
-      var defaultProperties = dataModel.entityForType(TotalModel).defaultProperties;
+      var defaultProperties =
+          dataModel.entityForType(TotalModel).defaultProperties;
       expect(defaultProperties.contains("id"), true);
       expect(defaultProperties.contains("field"), true);
       expect(defaultProperties.contains("addedField"), true);
 
-      expect(dataModel.entityForType(PartialReferenceModel).defaultProperties.contains("foreignKeyColumn"), true);
+      expect(
+          dataModel
+              .entityForType(PartialReferenceModel)
+              .defaultProperties
+              .contains("foreignKeyColumn"),
+          true);
     });
   });
 
-    test("Delete rule of setNull throws exception if property is not nullable",
+  test("Delete rule of setNull throws exception if property is not nullable",
       () {
     try {
       new ManagedDataModel([Owner, FailingChild]);
@@ -281,7 +315,10 @@ void main() {
       new ManagedDataModel([NoPrimaryKey]);
       expect(true, false);
     } on ManagedDataModelException catch (e) {
-      expect(e.message, contains("Class '_NoPrimaryKey' doesn't declare a primary key property"));
+      expect(
+          e.message,
+          contains(
+              "Class '_NoPrimaryKey' doesn't declare a primary key property"));
     }
   });
 
@@ -308,7 +345,10 @@ void main() {
       new ManagedDataModel([InvalidModel]);
       expect(true, false);
     } on ManagedDataModelException catch (e) {
-      expect(e.message, contains("Property 'uri' on '_InvalidModel' has an unsupported type"));
+      expect(
+          e.message,
+          contains(
+              "Property 'uri' on '_InvalidModel' has an unsupported type"));
     }
   });
 
@@ -318,7 +358,10 @@ void main() {
       new ManagedDataModel([InvalidTransientModel]);
       expect(true, false);
     } on ManagedDataModelException catch (e) {
-      expect(e.message, startsWith("Property 'uri' on '_InvalidTransientModel' has an unsupported type"));
+      expect(
+          e.message,
+          startsWith(
+              "Property 'uri' on '_InvalidTransientModel' has an unsupported type"));
     }
   });
 
@@ -327,7 +370,9 @@ void main() {
 
     setUp(() {
       dataModel = new ManagedDataModel([User, Item, Manager]);
-      ManagedContext.defaultContext = new ManagedContext(dataModel, new DefaultPersistentStore());;
+      ManagedContext.defaultContext =
+          new ManagedContext(dataModel, new DefaultPersistentStore());
+      ;
     });
 
     test("works for a data model", () {
@@ -532,6 +577,7 @@ class TotalModel extends ManagedObject<_TotalModel> implements _TotalModel {
   @managedTransientAttribute
   int transient;
 }
+
 class _TotalModel extends PartialModel {
   String addedField;
 }
@@ -550,18 +596,24 @@ class PartialModel {
   }
 }
 
-class PartialReferenceModel extends ManagedObject<_PartialReferenceModel> implements _PartialReferenceModel {}
+class PartialReferenceModel extends ManagedObject<_PartialReferenceModel>
+    implements _PartialReferenceModel {}
+
 class _PartialReferenceModel {
   @managedPrimaryKey
   int id;
 
   String field;
 
-  @ManagedRelationship.deferred(ManagedRelationshipDeleteRule.cascade, isRequired: true)
+  @ManagedRelationship.deferred(ManagedRelationshipDeleteRule.cascade,
+      isRequired: true)
   PartialModel foreignKeyColumn;
 }
 
-class DoubleRelationshipForeignKeyModel extends ManagedObject<_DoubleRelationshipForeignKeyModel> implements _DoubleRelationshipForeignKeyModel{}
+class DoubleRelationshipForeignKeyModel
+    extends ManagedObject<_DoubleRelationshipForeignKeyModel>
+    implements _DoubleRelationshipForeignKeyModel {}
+
 class _DoubleRelationshipForeignKeyModel {
   @managedPrimaryKey
   int id;
@@ -573,7 +625,10 @@ class _DoubleRelationshipForeignKeyModel {
   DoubleRelationshipHasModel isOneOf;
 }
 
-class DoubleRelationshipHasModel extends ManagedObject<_DoubleRelationshipHasModel> implements _DoubleRelationshipHasModel{}
+class DoubleRelationshipHasModel
+    extends ManagedObject<_DoubleRelationshipHasModel>
+    implements _DoubleRelationshipHasModel {}
+
 class _DoubleRelationshipHasModel {
   @managedPrimaryKey
   int id;