+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 1.0.5
 - Persistent types for `ManagedObject`s can now be subclasses of some other type, and inherited properties are also persistent properties.
-- `ManagedRelationship`s can have additional `partialManagedObject` metadata.
-    This allows `ManagedObject` subclasses to be declared in an imported package and still have relationships with `ManagedObject`s in the importing package.
+- `ManagedRelationship`s now have a `.deferred()` constructor. This allows `ManagedObject` subclasses to be declared in an imported package and still have relationships with `ManagedObject`s in the importing package.
 
 ## 1.0.4
 - BREAKING CHANGE: Added new `Response.contentType` property. Adding "Content-Type" to the headers of a `Response` no longer has any effect; use this property instead.